requirements.txt
keyboard
infi.systray
psutil
pycaw